+++ src/Modules/Tours/Explorer.Tours.Infrastructure/Database/Repositories/ReviewDatabaseRepository.cs
@@ -23,13 +23,14 @@ namespace Explorer.Tours.Infrastructure.Database.Repositories
             _dbSet = _dbContext.Set<Review>();
         }
 
-        public PagedResult<Review> GetPagedByTourId(int page, int pageSize, int tourId)
+        public PagedResult<Review> GetPagedByTourId(int page, int pageSize, long tourId)
         {
-            var task = _dbSet.Where(x => x.TourId == tourId).GetPagedById(page, pageSize);  
+            var task = _dbSet.Where(x => x.TourId == tourId).GetPagedById(page, pageSize);
+            //var task = _dbContext.Reviews.Include(r => r.Tour).GetPagedById(page, pageSize);
             task.Wait();
             return task.Result;
         }
-        public bool ReviewExists(long touristId, int tourId)
+        public bool ReviewExists(long touristId, long tourId)
         {
             var entity = _dbSet.Where(x => x.TouristId == touristId && x.TourId == tourId).FirstOrDefault();
             return entity != null;
